Move WFO5001 warning suppression to Program.cs
The <NoWarn> property group that suppressed the warning WFO5001
has been removed from the DarkModeApp.csproj file. A #pragma
warning disable WFO5001 directive has been added to the
Program.cs file to suppress the WFO5001 warning within that file.
This change limits the suppression to Program.cs instead of
applying it globally to the entire project.